requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
numpy>=1.24.0
tiktoken>=0.5.0
PyPDF2>=3.0.0
